Example directory refactor
